DY-35 : Clothe, Ajout des commentaires.

+++ src/main/java/fr/redteam/dressyourself/core/clothes/Clothe.java
@@ -4,7 +4,9 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
-
+/**
+ * This class define the concept of one clothe
+ */
 public class Clothe implements Serializable {
 
   private static final long serialVersionUID = 1L;
@@ -25,6 +27,9 @@ public class Clothe implements Serializable {
 
   private List<String> weather;
 
+  /**
+   * this a simple constructor with all argument are null except weather which is an empty arraylist
+   */
 
   public Clothe() {
     this.id = 0L;
@@ -37,6 +42,10 @@ public class Clothe implements Serializable {
     this.weather = new ArrayList<String>();
   }
 
+  /**
+   * this a simple constructor with a string for modelname all of others arguments are null except
+   * weather which is an empty arraylist
+   */
   public Clothe(String model) {
     this.model = model;
     this.weather = new ArrayList<String>();
@@ -48,6 +57,10 @@ public class Clothe implements Serializable {
     this.bodies = null;
   }
 
+  /**
+   * this a simple constructor with a string for modelname and the second string define the path all
+   * of others arguments are null except weather which is an empty arraylist
+   */
   public Clothe(String model, String image) {
     this.model = model;
     this.imageRelativePath = image;
@@ -59,68 +72,115 @@ public class Clothe implements Serializable {
     this.weather = new ArrayList<String>();
   }
 
+  /**
+   * get the clothe's id
+   */
   public long getId() {
     return id;
   }
 
+  /**
+   * Get the clothe's model name
+   */
   public String getModel() {
     return model;
   }
 
-  public void setId(int id) {
+  /**
+   * set the clothe's id with the long id
+   */
+  public void setId(long id) {
     this.id = id;
   }
 
+  /**
+   * Set the clothe's model name with the string model
+   */
   public void setModel(String model) {
     this.model = model;
   }
 
+  /**
+   * get the clothe's brand
+   */
   public String getBrand() {
     return brand;
   }
 
+  /**
+   * set the clothe's brand with the string brand
+   */
   public void setBrand(String brand) {
     this.brand = brand;
   }
 
+  /**
+   * get the clothe's color
+   */
   public String getColor() {
     return color;
   }
 
+  /**
+   * set the clothe's color with the string color
+   */
   public void setColor(String color) {
     this.color = color;
   }
 
+  /**
+   * get the clothe's path of clothe's image
+   */
   public String getImageRelativePath() {
     return imageRelativePath;
   }
 
+  /**
+   * set the clothe's path with the string imagePath
+   */
   public void setImageRelativePath(String imagePath) {
     this.imageRelativePath = imagePath;
   }
 
+  /**
+   * get the clothe's type
+   */
   public String getType() {
     return type;
   }
 
+  /**
+   * set the clothe's type with the string type
+   */
   public void setType(String type) {
     this.type = type;
   }
 
 
+  /**
+   * get the clothe's bodies
+   */
   public String getBodies() {
     return bodies;
   }
 
-
+  /**
+   * set the clothe's bodies with the string bodies
+   */
   public void setBodies(String bodies) {
     this.bodies = bodies;
   }
 
+  /**
+   * get the clothe's weather
+   */
   public List<String> getWeather() {
     return weather;
   }
 
+  /**
+   * set the clothe's weather with one List of string
+   */
   public void setWeather(List<String> weather) {
     this.weather = weather;
   }
